sampleapp/requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2.4
django-extensions==2.2.1
djangorestframework==3.10.2
factory-boy==2.12.0
Faker==2.0.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
mysql-connector-python==8.0.17
mysqlclient==1.4.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
